Setup and destroy header file working
