--- sql/database-test.cc
@@ -1,9 +1,0 @@
-//
-// Created by anuj on 24/09/20.
-//
-
-extern int mysqld_main(int argc, char **argv);
-
-int main(int argc, char **argv) {
-  return mysqld_main(argc, argv);
-}
